Rename SwiftPolyglot struct, add ParsableCommand adoption and add RuntimeError usage
